Fixed RGB LED implementation and Added Switch Buttons
The RGB LED on this project are common anode, this means that the power source will come from a PWM output, and we are going to use the digital OUTPUT in current sinking mode.
This means that i had to:
- Swapp HIGH - LOW digital OUTPUT signals. LOW now means it's gonna turn ON, and HIGH means it's gonna turn OFF.
- Add the PWM ports.

Also added the ports i, in particular, am gonna use in this project. Later i'll probably realease the house plant and the PCB that goes with the project.

Last thing i did was add the switch buttons. 
They work like this: 
- Turns ON white if LED is turned OFF.
- Turns OFF if any single LED (be it R, G or B, or all) is turned ON.