Another bunch of fixed stuff - working version
Fixed lot's of stuff by testing it and seeing what was going wrong. Swapped a lot of pins in the code to match the real life prototype and made a lot of other small changes. Schematics for the full project, to be printed in a PCB included, and circuits within explained.